engine: Remove redundant casting
Java 8 has considerably improved its ability to infer return values in
generic contexts, making many of the casts we had to perform in previous
versions redundant.

This patch removes these casts and cleans up the code accordingly.

+++ backend/manager/modules/bll/src/main/java/org/ovirt/engine/core/bll/scheduling/external/ExternalSchedulerBrokerImpl.java
@@ -35,8 +35,8 @@ public class ExternalSchedulerBrokerImpl implements ExternalSchedulerBroker {
         String extSchedUrl = Config.getValue(ConfigValues.ExternalSchedulerServiceURL);
         config = new XmlRpcClientConfigImpl();
         config.setEnabledForExtensions(true);
-        config.setConnectionTimeout((Integer) Config.getValue(ConfigValues.ExternalSchedulerConnectionTimeout));
-        config.setReplyTimeout((Integer) Config.getValue(ConfigValues.ExternalSchedulerResponseTimeout));
+        config.setConnectionTimeout(Config.getValue(ConfigValues.ExternalSchedulerConnectionTimeout));
+        config.setReplyTimeout(Config.getValue(ConfigValues.ExternalSchedulerResponseTimeout));
         try {
             config.setServerURL(new URL(extSchedUrl));
         } catch (MalformedURLException e) {
